+++ source/Knyaz.Optimus.Tests/Performance/ComputedStylePerformanceTest.cs
@@ -1,8 +1,5 @@
-﻿using System.IO;
-using System.Text;
-using Knyaz.Optimus.TestingTools;
+﻿using Knyaz.Optimus.TestingTools;
 using Knyaz.Optimus.Tests.TestingTools;
-using Knyaz.Optimus.Tests.Tools;
 using NBench;
 
 namespace Knyaz.Optimus.Tests.Performance
@@ -15,13 +12,6 @@ namespace Knyaz.Optimus.Tests.Performance
 		public void SetUp(BenchmarkContext ctx) => _counter = ctx.GetCounter("Counter");
 
 		
-		private Engine Load(string html)
-		{
-			var engine = TestingEngine.BuildJintCss();
-			engine.Load(new MemoryStream(Encoding.UTF8.GetBytes(html)));
-			return engine;
-		}
-		
 		[PerfBenchmark(NumberOfIterations = 3, RunTimeMilliseconds = 1000, RunMode = RunMode.Throughput)]
 		[CounterMeasurement("Counter")]
 		[MemoryMeasurement(MemoryMetric.TotalBytesAllocated)]
@@ -30,7 +20,10 @@ namespace Knyaz.Optimus.Tests.Performance
 			var styles = @"strong {font-family:""Arial""} .a string {font-family:""Curier New""}";
 			var html = "<div class a><span><strong id=test></strong><span></div>";
 			
-			var engine = Load("<head><style>" + styles + "</style></head><body>" + html + "</body>");
+			var resourceProvider = Mocks.ResourceProvider("http://localhost", 
+				"<head><style>" + styles + "</style></head><body>" + html + "</body>");
+
+			var engine = TestingEngine.BuildJintCss(resourceProvider);
 			var doc = engine.Document;
 			var elt = doc.GetElementById("test");
 			elt.GetComputedStyle().GetPropertyValue("font-family");